Subsume member_of_collections_display mixin into valkyrie/resource_decorator
